% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_estimation.R
\name{eval.thresh}
\alias{eval.thresh}
\title{Evaluate a high threshold of R|W at a set of new angles W}
\usage{
eval.thresh(fit.thresh.out, w.eval, up = 30)
}
\arguments{
\item{fit.thresh.out}{the output of the 'fit.thresh' function}

\item{w.eval}{a vector of length n or a matrix with n rows}

\item{up}{numerical value giving upper limit for root finding of r0w when using KDE method (default 30)}
}
\value{
vector containing elements r0w (estimated threshold for each given w.eval)
}
\description{
Evaluate a high threshold of R|W at a set of new angles W
}
