% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_solutions.R
\name{find_solutions}
\alias{find_solutions}
\title{Find the real solutions for x}
\usage{
find_solutions(gamma, theta, tol = 1e-09)
}
\arguments{
\item{gamma}{A numeric value in the interval (2^(-theta), 1].}

\item{theta}{A numeric value in the interval (0, 1].}

\item{tol}{The desired numerical tolerance for the solver. Default is 1e-9.}
}
\value{
A numeric vector containing the two solutions for x.
}
\description{
This function solves the equation(1-x^(1/theta))^theta - gamma = -((x^(1/theta-1))\emph{(1-x^(1/theta))^(theta-1))}(x-gamma)
for x, given the specified constraints on the parameters.
}
