% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodandmodelfitting_par.R
\name{fit.geometric.par}
\alias{fit.geometric.par}
\title{Fit truncated gamma model for R|W for bivariate data}
\usage{
fit.geometric.par(
  r,
  w,
  r0w,
  thresh.fit = NULL,
  model,
  customgauge = NULL,
  fixshape = FALSE,
  init.val = NULL,
  lower.limit = NULL,
  upper.limit = NULL,
  optimmethod = "Nelder-Mead",
  maxit = 1000,
  hessian = FALSE,
  nwg = 70,
  ...
)
}
\arguments{
\item{r}{Values of R|W=w that exceed the threshold r0w}

\item{w}{Values of W that correspond to the threshold exceedances of R}

\item{r0w}{The threshold function r_0(w) for each value of W supplied}

\item{thresh.fit}{The output from fit.thresh. If this is included, then exceedances of r,w, and the corresponding threshold r0w, are automatically calculated, and they overwrite any supplied r, w, r0w}

\item{model}{Character string specifying model to use for the gauge function. Options are "log" for logistic-type gauge, "invlog" for inverted logistic, "gauss" for Gaussian, "square" for square, "invclayton" for inverted Clayton, "expgauss", "expinvlog","expsquare" for Gaussian, inverted logistic or square mixed with an expenential variable (see Lee and Wadsworth (2025)), and "alogistic" for asymmetric logistic. If two options are given, then the fitted gauge will be an additive mixture of these two.}

\item{customgauge}{Optional argument to specify a custom gauge function. This should have}

\item{fixshape}{Logical. If TRUE, then the shape parameter of the truncated gamma distribution is fixed to the dimension d, otherwise this is estimated.}

\item{init.val}{vector of initial gauge function parameter values for optimization (needed when customgauge used, otherwise optional)}

\item{lower.limit}{vector of lower limits on parameters (when customgauge used; should be same length as init.val)}

\item{upper.limit}{vector of upper limits on parameters (when customgauge used; should be same length as init.val)}

\item{nwg}{number of points used in expanding mesh to cover simplex for numerically rescaled additive gauge}

\item{...}{Additional arguments to pass to the gauge function (if necessary)}
}
\value{
list containing elements "mle", "nllh", "convergence", and additionally "cov" if argument hessian=T is supplied to optim
}
\description{
Fit truncated gamma model for R|W for bivariate data
}
