% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodandmodelfitting_pwl.R
\name{fit.geometric.pwl}
\alias{fit.geometric.pwl}
\title{Fit piecewise-linear truncated gamma model for R|W, and a piecewise-linear model for W}
\usage{
fit.geometric.pwl(
  r,
  w,
  r0w,
  locs = NULL,
  init.val = NULL,
  fixshape = FALSE,
  W.fit = FALSE,
  joint.fit = FALSE,
  method = "BFGS",
  pen.const = 1,
  bound.fit = FALSE,
  hessian = FALSE
)
}
\arguments{
\item{r}{Values of R|W=w that exceed the threshold r0w}

\item{w}{Values of W that correspond to the threshold exceedances of R}

\item{locs}{The reference angles (nodes) at which parameters are estimation}

\item{init.val}{vector of initial parameter values for optimization (when customgauge used)}

\item{fixshape}{Logical. If TRUE, then the shape parameter of the truncated gamma distribution is fixed to the dimension d, otherwise this is estimated.}

\item{W.fit}{Logical. If TRUE, then the angular model is fitted.}

\item{joint.fit}{the optimization method (default is "BFGS")}

\item{pen.const}{the L2 penalty strength (default is 1)}

\item{bount.fit}{Logical. If TRUE, the gauge function is estimated such that its unit level set has coordinatewise maximum 1}
}
\value{
list containing elements "mle", "nllh", "convergence", and additionally "cov" if argument hessian=T is supplied to optim
}
\description{
Fit piecewise-linear truncated gamma model for R|W, and a piecewise-linear model for W
}
