% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_new.R
\name{iweights}
\alias{iweights}
\title{Calculate importance weights for simulation of W|R'>k, k>1, (equivalently W|R>k r_0(W), where R'=R/r_0(W)).}
\usage{
iweights(
  k,
  r0w,
  w,
  gfun,
  par,
  add.gauge = FALSE,
  gauge1,
  gauge2,
  par1ind,
  par2ind,
  shape,
  nwg
)
}
\arguments{
\item{k}{value of k>1: the importance weights are to simulate from the distribution of W|R'>k}

\item{r0w}{The threshold function r_0(w) for each value of W supplied}

\item{w}{Values of W that correspond to the threshold exceedances of R}

\item{gfun}{fitted gauge function}

\item{par}{fitted parameter values (first value should be the gamma shape parameter, subsequent values are the gauge function parameters)}

\item{add.gauge}{logical. If TRUE, then we additively mix (with numerical rescaling) the gauge functions specified by gauge1 and gauge2}

\item{gauge1}{first gauge function for additive mixing. Use if add.gauge=TRUE}

\item{gauge2}{second gauge function for additive mixing. Use if add.gauge=TRUE}

\item{par1ind}{vector of indices specifying the position of the parameters for gauge1 in par}

\item{par2ind}{vector of indices specifying the position of the parameters for gauge2 in par}

\item{nwg}{number of points to use in expanding wgrid for numerical rescaling of additive mixture in >2 dimensions}
}
\value{
vector of (unnormalized) importance weights to be used in sim.2d
}
\description{
Calculate importance weights for simulation of W|R'>k, k>1, (equivalently W|R>k r_0(W), where R'=R/r_0(W)).
}
