% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeloptions.R
\name{modeloptions}
\alias{modeloptions}
\title{Details model options for parametric fitting (to be used internally)}
\usage{
modeloptions(
  model,
  customgauge,
  init.val,
  lower.limit,
  upper.limit,
  d,
  theta12,
  theta13,
  theta23,
  theta123
)
}
\arguments{
\item{model}{Character string specifying which gauge model to use. If two names are provided, these are additively mixed with a weight parameter.}

\item{customgauge}{Optional custom gauge function}

\item{init.val}{initial values of custom gauge function parameters (must be supplied if customgauge used)}

\item{lower.limit}{lower bounds of custom gauge function parameters (must be supplied if customgauge used)}

\item{upper.limit}{upper bounds of custom gauge function parameters (must be supplied if customgauge used)}

\item{d}{dimension of model fit}

\item{theta12}{used if moel = "alogistic" to specify whether there are joint extremes for this group}

\item{theta13}{used if moel = "alogistic" to specify whether there are joint extremes for this group}

\item{theta23}{used if moel = "alogistic" to specify whether there are joint extremes for this group}

\item{theta123}{used if moel = "alogistic" to specify whether there are joint extremes for this group}
}
\description{
Details model options for parametric fitting (to be used internally)
}
