% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_new.R
\name{plotfittedgauge}
\alias{plotfittedgauge}
\title{Plot fitted limit set (unit level set of gauge function) for d=2,3}
\usage{
plotfittedgauge(
  fit,
  resolution = 70,
  add = FALSE,
  col = "red",
  xlab = "x",
  ylab = "y",
  zlab = "z",
  lwd = 2,
  type = "l",
  polygon = TRUE,
  polygoncol = rgb(0.5, 0, 0, alpha = 0.5),
  alphaval = 0.5,
  ...
)
}
\arguments{
\item{fit}{list output from fit.geometric.par or fit.geometric.pwlin}

\item{resolution}{plotting resolution - the higher the number, the finer the resolution of the plot. For d=2, the default should suffice, but larger numbers should not take too much longer to plot. For d=3, higher numbers will normally take longer to plot, but give better figures.}

\item{add}{logical - add plot to an existing figure?}

\item{col}{character string or number giving the colour for the plot}

\item{xlab, ylab, zlab}{character string giving axis labels}

\item{lwd, type}{line width and type arguments to plot for d=2}

\item{polygon}{logical - include shaded polygon of limit set for d=2?}

\item{polygoncol}{character string or number giving the colour for the polygon}

\item{alphaval}{alpha value for shading of the polygon or 3d level set}

\item{...}{additional arguments to pass to plot for d=2}
}
\value{
plot of fitted gauge function
}
\description{
Plot fitted limit set (unit level set of gauge function) for d=2,3
}
