% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_new.R
\name{plotfittedthresh}
\alias{plotfittedthresh}
\title{Plot fitted threshold for d=2,3}
\usage{
plotfittedthresh(thresh.fit, resolution = 100, add = FALSE)
}
\arguments{
\item{thresh.fit}{output from fit.thresh}

\item{resolution}{number giving plotting resolution of fitted threshold. Higher numbers plot at more points, but take longer (particularly for d=3)}

\item{add}{logical - add to existing plot?}
}
\value{
plot of fitted threshold over data sample
}
\description{
Plot fitted threshold for d=2,3
}
