% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_new.R
\name{ppdiag}
\alias{ppdiag}
\title{PP diagnostic for fitted truncated gamma model}
\usage{
ppdiag(fit)
}
\arguments{
\item{fit}{list output from fit.geometric.par or fit.geometric.pwlin}
}
\value{
PP plot with 95\\% tolerance intervals
}
\description{
PP diagnostic for fitted truncated gamma model
}
