% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_new.R
\name{qqdiag}
\alias{qqdiag}
\title{QQ diagnostic for fitted truncated gamma model - puts PP plot on any desired quantile scale (default is standard exponential)}
\usage{
qqdiag(fit, quantilefn = qexp)
}
\arguments{
\item{fit}{list output from fit.geometric.par or fit.geometric.pwlin}

\item{quantilefn}{a quantile function to specify the desired scale. Defaults to qexp.}
}
\value{
QQ plot with 95\\% tolerance intervals
}
\description{
QQ diagnostic for fitted truncated gamma model - puts PP plot on any desired quantile scale (default is standard exponential)
}
