% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_new.R
\name{Rexc.prob.k}
\alias{Rexc.prob.k}
\title{Calculate P(R'>k|R'>1) (equivalently P(R>kr_0(W)|R>r_0(w)))}
\usage{
Rexc.prob.k(k, fit, nwg = 70)
}
\arguments{
\item{k}{value of k>1}

\item{fit}{list output from fit.geometric.par or fit.geometric.pwlin}

\item{nwg}{value giving resolution for numerical rescaling of additive gauges (where needed). Larger numbers give greater accuracy, but take longer.}
}
\value{
Estimate of P(R'>k|R'>1)
}
\description{
Calculate P(R'>k|R'>1) (equivalently P(R>kr_0(W)|R>r_0(w)))
}
