% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodandmodelfitting_par.R
\name{rw.gamma.lik}
\alias{rw.gamma.lik}
\title{Negative log likelihood function for truncated gamma model for R|W}
\usage{
rw.gamma.lik(
  psi,
  r,
  w,
  r0w,
  gfun,
  add.gauge = FALSE,
  gauge1,
  gauge2,
  par1ind,
  par2ind,
  lower.limit,
  upper.limit,
  d,
  wgrid = NULL,
  fixshape,
  ...
)
}
\arguments{
\item{psi}{vector of parameters. First component is the gamma shape parameter}

\item{r}{Values of R|W=w that exceed the threshold r0w}

\item{w}{Values of W that correspond to the threshold exceedances of R}

\item{r0w}{The threshold function r_0(w) for each value of W supplied}

\item{gfun}{gauge function to be used in the likelihood. Can use inbuilt examples or specify own function of the form gauge<-function(xy,par){}. Leave blank if want to mix gauges additively.}

\item{add.gauge}{logical. If TRUE, then we additively mix (with numerical rescaling) the gauge functions specified by gauge1 and gauge2}

\item{gauge1}{first gauge function for additive mixing. Use if add.gauge=TRUE}

\item{gauge2}{second gauge function for additive mixing. Use if add.gauge=TRUE}

\item{par1ind}{vector of indices specifying the position of the parameters for gauge1 in psi}

\item{par2ind}{vector of indices specifying the position of the parameters for gauge2 in psi}

\item{lower.limit}{optional vector of lower limits on parameters (should be same length as psi)}

\item{upper.limit}{optional vector of upper limits on parameters (should be same length as psi)}

\item{d}{dimension}

\item{wgrid}{matrix containing points that cover the d-1 simplex. Used only if add.gauge=T, to perform numerical rescaling}

\item{fixshape}{Logical. If TRUE, then the shape parameter of the truncated gamma distribution is fixed to the dimension d, otherwise this is estimated.}

\item{pos.par}{logical. Set to TRUE if all parameters are positive (avoids some warnings in optimization)}
}
\value{
negative log-likelihood value
}
\description{
Negative log likelihood function for truncated gamma model for R|W
}
