% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_new.R
\name{sim.geometric}
\alias{sim.geometric}
\title{Simulate new data (using resampled W|R'>k) from the model, conditional on R'>k (R>k*r_0(W))}
\usage{
sim.geometric(fit, nsim, k = 1, nwg = 70)
}
\arguments{
\item{fit}{list output from fit.geometric.par or fit.geometric.pwlin}

\item{nsim}{number of new datapoints to simulate}

\item{k}{Value of k (defaults to 1). If k>1 then importance weights are calculated to simulate from the distribution of X|R>kr_0(W) (equivalently X|R'>k)}

\item{nwg}{number of points to use in expanding wgrid for numerical rescaling of additive mixture in >2 dimensions}
}
\value{
nsim by 2 matrix of values (in exponential margins) simulated from the model (given R>r_0(W), or R>kr_0(W) for k>1)
}
\description{
Simulate new data (using resampled W|R'>k) from the model, conditional on R'>k (R>k*r_0(W))
}
