% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_new.R
\name{tail.dep.plots}
\alias{tail.dep.plots}
\title{Create chi(u) plots}
\usage{
\method{tail}{dep.plots}(thresh.fit, fit, dataset)
}
\arguments{
\item{thresh.fit}{output from fit.thresh}

\item{fit}{list output from fit.geometric.par or fit.geometric.pwlin}

\item{dataset}{dataset x in exponential margins}
}
\value{
plots of chi(u)
}
\description{
Create chi(u) plots
}
