% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_estimation.R
\name{fit.thresh}
\alias{fit.thresh}
\title{Find a high threshold of R|W which is (approximately) the tau-quantile of this variable}
\usage{
fit.thresh(r, w, tau = 0.95, method = c("empirical", "KDE"))
}
\arguments{
\item{r}{Values of radial variable}

\item{w}{values of angular variable (on unit simplex)}

\item{tau}{level at which to calculate threshold}

\item{method}{character string specifying either "empirical" for using a binning mathod, or "KDE" for the method based on kernel density estimation as outlines in Campbell and Wadsworth (2025)}
}
\value{
list containing elements r0w (estimated threshold for each given w), r, w, tau and method
}
\description{
Find a high threshold of R|W which is (approximately) the tau-quantile of this variable
}
